SF: converting mixed Priority Realtime definitons Commit 0c357a111d3e5bf3a46a51472a4acd2497072dfe was relying on a mix of PRIORITY_URGENT_DISPLAY and PRIORITY_REALTIME. That results in a huge nice value, which makes the task seem important.
However that was missing the real purpose of
the PRIORITY_REALTIME definition.
By summing the values of the two definitions
the Threads controller wasn't able to recognize
ANDROID_PRIORITY_REALTIME which is hardcoded at -21
and apply the important SP_REALTIME scheduling policy
on it.

This commit fixes the flaws by relying on the
PRIORITY_REALTIME definition only.